+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 
 ### Added
 
+- `limestone/Alert` `buttonDirection` prop to control button layout direction (`auto`, `horizontal`, `vertical`)
 - `limestone/Icon` supported icon list, adding new icon `f1`
 - `limestone/Input` props `marqueeInputField` to marquee inputField text on focus
 - `limestone/Inputfield` prop `caretToEndOnFocus` to move caret to the end of the input field on input activation via 5-way
@@ -13,6 +14,7 @@ The following is a curated list of changes in the Enact limestone module, newest
 - `limestone/PageViews` `showFooterButtons` prop to show optional Close/Next buttons
 - `limestone/PageViews` `footerCloseLabel` prop to custom the Close button label
 - `limestone/PageViews` `onFooterCloseClick` prop to custom the Close button behavior when it is clicked
+- `limestone/Slider` `colorPicker` prop to change slider functionality
 - `limestone/ThemeDecorator` config `focusEffectClass` to support customization for spottable components
 
 ### Changed